rename prompt folder
