+++ app.py
@@ -7,7 +7,7 @@ from flask import Flask, request
 from pymessenger.bot import Bot
 
 import database.db_test_scripts as db
-from src.prompt_design import gen_response
+from promt.prompt_design import gen_response
 
 app = Flask(__name__)
